requirements.txt
Django==4.0.\5
djangorestframework==3.13.1
markdown==3.3.7
django-filter==21.1
djangorestframework-simplejwt==5.2.0
pillow==9.1.1